idle-monitor: Make sure we publicly make watch intervals 64-bits wide
https://bugzilla.gnome.org/show_bug.cgi?id=688227

+++ libgnome-desktop/gnome-idle-monitor.c
@@ -470,7 +470,7 @@ gnome_idle_monitor_new_for_device (GdkDevice *device)
  */
 guint
 gnome_idle_monitor_add_watch (GnomeIdleMonitor	       *monitor,
-			      guint			interval_msec,
+			      guint64			interval_msec,
 			      GnomeIdleMonitorWatchFunc callback,
 			      gpointer			user_data,
 			      GDestroyNotify		notify)
